### 2018-04-12 70cf42a
doc : paper modified
