+++ paper/paper.md
@@ -46,7 +46,9 @@ With a valid mathematical model, PEMFC system performance can be better understo
 ![Block diagram of software](../otherfile/OPEM_BLOCK_DIAGRAM.jpg) 
 
 
-Supported models include two general categories of static models and dynamic models. Static models are briefly outlined here:
+Supported models include two general categories of static models and dynamic models. Static models are briefly outlined here:							
+
+
 ### Amphlett analysis
 that main concepts include Nernst voltage, PEMFC losses (activation polarization loss, ohmic polarization loss, and concentration polarization loss), power and efficiency of the fuel cell.[@Amphlett1995; @Correa2005; @Saadi2013]
 
@@ -57,7 +59,9 @@ that was developed to proportionate the experimental cell potential (E) vs. curr
 
 that is represented by means of its voltage-current characteristic obtained in static operating mode. In fact, Larminie-Dicks static model presents the fuel cell voltage as a function of the current magnitude.[@Kim1995; @Saadi2013]
 
-And also dynamic models includes:
+And also dynamic models includes:								
+
+
 ### Padulles analysis I,II
 
 that the electrodes channel temperatures/pressures, transient response of cell voltage, gases outflow rates and the temperature of the cell under the sudden change in load current can be predicted. In model I, Nernst and fuel cell potential were modeled as a function of oxygen and hydrogen gases partial pressure that can be calculated from independent variables or constants and in model II, Nernst and fuel cell potential were modeled as a function of water, oxygen and hydrogen gases partial pressure that can be calculated from independent variables or constants.[@Padulls2000]
