### 2018-04-12 1011c8c
fix : paper Hauer2001 reference fixed
