+++ paper/paper.bib
@@ -66,7 +66,7 @@
   journal = {Journal of Power Sources}
 }
 
-@techreport{hauer2001analysis,
+@techreport{Hauer2001,
   title={Analysis tool for fuel cell vehicle hardware and software (controls) with an application to fuel economy comparisons of alternative system designs},
   author={Hauer, Karl-Heinz},
   year={2001}
