### 2018-04-12 50ace3b
doc : supported models section added
