### 2018-04-14 ba9da21
doc : minor edit in JOSS paper
