### 2018-04-15 #672 @kyleniemeyer
@danielskatz Practically I don't think I'd be able to handle this within the next week. @katyhuff are you available right now?
