### 2018-04-15 #672 @katyhuff
I'm swamped this week, too, but this does look like a strong submission and hopefully will be fairly straightforward. I can take it.
