### 2018-04-15 #676 @whedon
[REVIEW]: OPEM (Open Source PEM Fuel Cell Simulation Tool)
**Submitting author:** @sepandhaghighi (<a href="http://orcid.org/0000-0001-9450-2375">Sepand Haghighi</a>)
**Repository:** <a href="https://github.com/ECSIM/opem" target ="_blank">https://github.com/ECSIM/opem</a>
**Version:** v0.8
