### 2018-04-15 #672 @katyhuff
Thanks folks. @nnadeau and @sepandhaghighi , please let's head over to #676 for the review itself.
