+++ README.md
@@ -37,7 +37,7 @@
 ## Overview		
 
 <p align="justify">	
-Modelling and simulation of proton-exchange membrane fuel cells (PEMFC) may work as a powerful tool in the Research & development of renewable energy sources.The Open-Source PEMFC Simulation Tool (OPEM) is a modelling tool for evaluating the performance of proton exchange membrane fuel cells. This package is a combination of some models (static/dynamic) which are predicting the optimum operating parameters of PEMFC. OPEM contained generic models that will accept as input, not only values of the operating variables such as anode and cathode feed gas, pressure and compositions, cell temperature and current density, but also cell parameters including the active area and membrane thickness. In addition, some of the different models of PEMFC that have been proposed in the OPEM, just focus on one particular FC stack,and some others take into account a part or all auxiliaries such as reformers. OPEM is a platform for collaborative development of PEMFC models.
+Modelling and simulation of proton-exchange membrane fuel cells (PEMFC) may work as a powerful tool in the research & development of renewable energy sources.The Open-Source PEMFC Simulation Tool (OPEM) is a modelling tool for evaluating the performance of proton exchange membrane fuel cells. This package is a combination of some models (static/dynamic) which are predicting the optimum operating parameters of PEMFC. OPEM contained generic models that will accept as input, not only values of the operating variables such as anode and cathode feed gas, pressure and compositions, cell temperature and current density, but also cell parameters including the active area and membrane thickness. In addition, some of the different models of PEMFC that have been proposed in the OPEM, just focus on one particular FC stack,and some others take into account a part or all auxiliaries such as reformers. OPEM is a platform for collaborative development of PEMFC models.
 </p>
 <div align="center">
 
