### 2018-04-19 ECSIM/opem#34 @nnadeau
`check-manifest`: List of files in version control and sdist do not match
- Given PyPi's [`check-manifest`](https://pypi.org/project/check-manifest/) tool:

```
$ check-manifest
```

```
lists of files in version control and sdist do not match!
missing from sdist:
  AUTHORS.md
  CHANGELOG.md
  CONTRIBUTING.md
  Documents
  Documents/Amphlett.ipynb
  Documents/Chamberline_Kim.ipynb
  Documents/Larminie_Dicks.ipynb
  Documents/Padulles1.ipynb
  Documents/Padulles2.ipynb
  Documents/Padulles_Amphlett.ipynb
  Documents/Padulles_Hauer.ipynb
  Documents/README.md
  Documents/images
  Documents/images/Padulles-Amphlett-Diagram.jpg
  Documents/images/Padulles-Diagram.jpg
  Documents/images/Padulles-Hauer-Diagram.jpg
  Documents/images/Padulles2-Diagram.jpg
  Documents/images/Static-Model-Graphic.jpg
  INSTALL.md
  OPEM.spec
  PEM.md
  appveyor.yml
  otherfile
  otherfile/Fuel Cell Graphic.gif
  otherfile/OPEM_BLOCK_DIAGRAM.jpg
  otherfile/PEMFC1.jpg
  otherfile/PEMFC2.jpg
  otherfile/Version.rc
  otherfile/icon.ico
  otherfile/logo.png
  otherfile/test.csv
  otherfile/test.html
  otherfile/test.opem
  paper
  paper/paper.bib
  paper/paper.md
  requirements.txt
suggested MANIFEST.in rules:
  include *.md
  include *.spec
  include *.txt
  include *.yml
  recursive-include Documents *.ipynb
  recursive-include Documents *.jpg
  recursive-include Documents *.md
  recursive-include otherfile *.csv
  recursive-include otherfile *.gif
  recursive-include otherfile *.html
  recursive-include otherfile *.ico
  recursive-include otherfile *.jpg
  recursive-include otherfile *.opem
  recursive-include otherfile *.png
  recursive-include otherfile *.rc
  recursive-include paper *.bib
  recursive-include paper *.md
```

- Files and directories should be explicitly included/excluded/pruned from `MANIFEST.ini`
