### 2018-04-19 ECSIM/opem#35 @nnadeau
Nitpick: Modelling vs. Modeling
- Modelling/Modeling used interchangeably throughout documentation and source code
- Pick one:
    - Modelling = international
    - Modeling = US
