### 2018-04-19 ECSIM/opem#36 @nnadeau
Avoid `import *`
- Especially in examples
  - E.g.,: https://github.com/ECSIM/opem/blob/master/Documents/Amphlett.ipynb
    - It is not obvious where `xi1`, `xi2`, etc come from
- Most linters will flag from module import * with a warning, because it's prone to namespace collisions
