### 2018-04-19 ECSIM/opem#37 @nnadeau
Follow PEP8 style guidelines (or explicitly define your own)
- It would be preferable if the package followed [PEP8](https://www.python.org/dev/peps/pep-0008/) guidelines
- From the PEP8 documentation:
```
One of Guido's key insights is that code is read much more often than it is written. The guidelines provided here are intended to improve the readability of code and make it consistent across the wide spectrum of Python code. As PEP 20 says, "Readability counts".

A style guide is about consistency. Consistency with this style guide is important. Consistency within a project is more important. Consistency within one module or function is the most important.
```
