### 2018-04-19 ECSIM/opem#38 @nnadeau
Test against `flake8`
- It would be preferable if the package was tested against [`flake8`](http://flake8.pycqa.org/en/latest/)
- Linting helps maintain a robust and healthy code-base
- Related issue: https://github.com/ECSIM/opem/issues/37
- References:
  - https://medium.com/python-pandemonium/what-is-flake8-and-why-we-should-use-it-b89bd78073f2
  - https://simpleisbetterthancomplex.com/packages/2016/08/05/flake8.html
