### 2018-04-19 ECSIM/opem#39 @nnadeau
Requirements should be pinned
- `opem/requirements.txt`
- Look into https://pyup.io for automated dependency management
- Considering this is an application and not an extensible library, the requirements should be pinned to specific versions in order to guarantee reproducibility.
- Unpinned requirements can lead to sudden CI failures.
- See the following links for best practice:
    - https://packaging.python.org/discussions/install-requires-vs-requirements/
    - https://caremad.io/posts/2013/07/setup-vs-requirement/
    - https://docs.pipenv.org/
