### 2018-04-19 ECSIM/opem#40 @nnadeau
Do not use `except Exception`
- E.g.,: https://github.com/ECSIM/opem/blob/d9748c65ff00ff9f6b730420dc7bb6cad8bc2fca/opem/Static/Amphlett.py#L22-L25
- See the following links for why this is considered poor coding practice:
    - https://stackoverflow.com/questions/10594113/bad-idea-to-catch-all-exceptions-in-python
    - https://stackoverflow.com/questions/2416316/why-is-the-catchexception-almost-always-a-bad-idea
    - https://softwareengineering.stackexchange.com/questions/164256/is-catching-general-exceptions-really-a-bad-thing
- If this is really necessary, the reasoning should be commented.
