### 2018-04-19 ECSIM/opem#41 @nnadeau
Duplicate/redundant requirements definition
- Requirements are doubly defined in both

https://github.com/ECSIM/opem/blob/ba9da21f9225df4df01171f9d2c47c7ee1e8765c/setup.py#L29-L33

https://github.com/ECSIM/opem/blob/ba9da21f9225df4df01171f9d2c47c7ee1e8765c/requirements.txt#L1-L3

- This can lead to out-of-sync code
- See links in https://github.com/ECSIM/opem/issues/39 for more information on good practice
