### 2018-04-19 ECSIM/opem#42 @nnadeau
General syntax/grammar fixes in `opem/Documents/Amphlett.ipynb`
- There are several grammar/syntax errors throughout the example that reduce the readability of the document.
- E.g.,:
  - `is a parametric model that predicting the performance`
  - ` Main concepts in the Amphlett model includes Nernst voltage, PEMFC`
  - `This parametric model of PEMFC using a combination of mechanistic and empirical approach.`
  - `For most fuel cells and their practical operating range, a linear approximation`
