### 2018-04-19 ECSIM/opem#43 @nnadeau
General grammar/syntax in `Documents/Larminie_Dicks.ipynb`
E.g.,:
- `or most fuel cells and their practical operating range, a linear approximation is actually a very good fit`
