### 2018-04-19 ECSIM/opem#44 @nnadeau
General grammar/syntax fixes in `Documents/Padulles1.ipynb`, `Documents/Padulles2.ipynb`, `Documents/Padulles_Amphlett.ipynb`, and `Documents/Padulles_Hauer.ipynb`
E.g.,:
- `This model based on some assumption such; the gases are ideal,`
- `..., The channels that transport gases along the electrodes have a fixed volume, only source of losses is ohmic and Nernst equation can be applied too.`
- `For most fuel cells and their practical operating range, a linear approximation is actually a very good fit`
