### 2018-04-19 ECSIM/opem#45 @nnadeau
General grammar/syntax fixes in paper
E.g.,:
- ` In addition to the decreasing fuel sources, also pollution and global ...`
- ` reaction between hydrogen-rich fuel source and oxygen`
- `... of these devices owing to high efficiency relative`
- ` PEM fuel cells are a high-efficiency power source without pollution for urban vehicles, corroborated by recent legislative initiatives, That’s why these devices have become a great deal of attention to the option of fuel-cell-powered vehicles`
- `Best fuel for fuel cell powered vehicle is hydrogen,`
- `because the water is an only product of the reaction between hydrogen and air`
- `protons that, flow via the membrane to the cathode`
- `An external circuit that linked to the two electrodes containing the collection of electrons as electrical current`
- ` water and heat will be produced from the protons that flowing via the electrolyte membrane`
- `The OPEM (software named based on open proton exchange membrane) written in Python, Python is a language that is very powerful for developers, but is also accessible to scientists.`
- `In addition, dynamic models improve static models and complete the simulation process.`
- `fluid science and material merge`
- ` make the simulation of fuel cell more `
- ` Simple linear regression used for this approximation and useful overall parameters such P(Thermal) and P(Elec) calculated by Simpson’s Rule`
- `The method for using in this software is the specified inputs are presented according to
the equations of each model by the user and outputs is in `
- `That is represented by means of its voltage-current characteristic obtained in static operating
mode`
- `That the electrodes channel temperatures/pressures, t...`
- `variables or constants and in model II, N...` (split sentence)

- Also, watch out for arbitrary capitalization, e.g.,: `hydrogen` vs `Hydrogen`
