### 2018-04-19 ECSIM/opem#46 @nnadeau
Reference past research
The statement `Previously, some research on modeling and simulation of PEMFC has been performed` is not backed with references.
