### 2018-04-19 ECSIM/opem#47 @nnadeau
Plagiarism of `PEM.md`
In its current form, `PEM.md` is plagiarized, word-for-word, from http://physics.oregonstate.edu/~hetheriw/energy/topics/doc/electrochemistry/fc/basic/The_Polymer_Electrolyte_Fuel_Cell.htm
