### 2018-04-19 ECSIM/opem#47 @nnadeau
Unfortunately, https://github.com/ECSIM/opem/issues/47 appears to be a case of plagiarism in the documentation. Considering that the entirety of `PEM.md` is copied, permission should be granted from the original author. Otherwise the reader should be referred directly to the original source.
