### 2018-04-19 #676 @nnadeau
Hi @sepandhaghighi and @katyhuff,

After performing an initial review, I have submitted the following issues:

## Plagarism
Unfortunately, https://github.com/ECSIM/opem/issues/47 appears to be a case of plagiarism in the documentation. Considering that the entirety of `PEM.md` is copied, permission should be granted from the original author. Otherwise the reader should be referred directly to the original source.

## Paper
- https://github.com/ECSIM/opem/issues/46

## General syntax/grammer
The following issues reduce the readability of the documentation, examples, and paper:
- https://github.com/ECSIM/opem/issues/45
- https://github.com/ECSIM/opem/issues/44
- https://github.com/ECSIM/opem/issues/43
- https://github.com/ECSIM/opem/issues/42
- https://github.com/ECSIM/opem/issues/35

## Code Quality
The following issues describe desired improvements to the codebase:
- https://github.com/ECSIM/opem/issues/41
- https://github.com/ECSIM/opem/issues/40
- https://github.com/ECSIM/opem/issues/39
- https://github.com/ECSIM/opem/issues/38
- https://github.com/ECSIM/opem/issues/37
- https://github.com/ECSIM/opem/issues/36
- https://github.com/ECSIM/opem/issues/34
