### 2018-04-19 ed8d30a
doc : Modelling renamed to Modeling in README file fixed #35
