### 2018-04-19 e1942b7
doc : Modelling renamed to Modeling in params.py fixed #35
