+++ opem/Params.py
@@ -3,17 +3,17 @@ Version=0.8
 Website="http://www.ecsim.ir/opem"
 UpdateUrl="http://www.ecsim.ir/opem/update"
 Overview='''
-Modelling and simulation of proton-exchange membrane fuel cells (PEMFC) may work as a
+Modeling and simulation of proton-exchange membrane fuel cells (PEMFC) may work as a
 powerful tool in the Research & development of renewable energy sources. The Open-Source
-PEMFC Simulation Tool (OPEM) is a modelling tool for evaluating the performance of proton
-exchange membrane fuel cells. This package is a combination of some models (static/dynamic)
-which are predicting the optimum operating parameters of PEMFC. OPEM contained generic
-models that will accept as input, not only values of the operating variables such as anode and
-cathode feed gas, pressure and compositions, cell temperature and current density, but also cell
+PEMFC Simulation Tool (OPEM) is a modeling tool for evaluating the performance of proton
+exchange membrane fuel cells. This package is a combination of models (static/dynamic)
+that predict the optimum operating parameters of PEMFC. OPEM contained generic models that
+will accept as input, not only values of the operating variables such as anode and cathode
+feed gas, pressure and compositions, cell temperature and current density, but also cell
 parameters including the active area and membrane thickness. In addition, some of the different
-models of PEMFC that have been proposed in the OPEM, just focus on one particular FC stack, and
-some others take into account a part or all auxiliaries such as reformers. OPEM is a platform for
-collaborative development of PEMFC models.'''
+models of PEMFC that have been proposed in the OPEM, just focus on one particular FC stack,
+and some others take into account a part or all auxiliaries such as reformers. OPEM is
+a platform for collaborative development of PEMFC models.'''
 
 Warning_Message_1="Warning : The value of I(>{}) leads to minus amount of V, please check your inputs"
 Warning_Message_2="Warning : There are errors in the simulations in some of I amounts; please refer to the .opem file" \
