### 2018-04-19 8d0fad4
doc : Reference for past research added to paper fixed #46
