+++ paper/paper.md
@@ -39,7 +39,7 @@ bibliography: paper.bib
 						
 
 # Summary
-The sun and wind as renewable energy sources‏ are attracting more regard as alternative energy sources. In addition to the decreasing fuel sources, also pollution and global warming‏ are important problems. Fuel cells are a beneficial energy technology that generates electric energy through the reaction between hydrogen-rich fuel source and oxygen. The advantage of these devices owing to high efficiency relative to combustion engines, low emissions and producing only heat and water as waste products. Proton exchange membrane fuel cells (PEMFCs) have attracted much interest recently. PEM fuel cells are a high-efficiency power source without pollution for urban vehicles, corroborated by recent legislative initiatives, That's why these devices have become a great deal of attention to the option of fuel-cell-powered vehicles. Previously, some research on modeling and simulation of PEMFC has been performed. The fundamental structure of a PEMFC can be described as two electrodes (anode and cathode) separated by a solid polymer membrane that acts as an electrolyte. Best fuel for fuel cell powered vehicle is hydrogen, this gas causing the highest conversion efficiency for fuel and generating zero tail-pipe emission because the water is an only product of the reaction between hydrogen and air. By flowing Hydrogen fuel through a network of channels to the anode, hydrogen separates into protons that, flow via the membrane to the cathode. An external circuit that linked to the two electrodes containing the collection of electrons as electrical current. Through a similar network of channels the oxygen that comes from the air, named oxidant, flows to the cathode and then, the electrons coming from the external electrical circuit will be received by oxygen and finally, water and heat will be produced from the protons that flowing via the electrolyte membrane.[@Feroldi2011; @Gottesfeld1999]
+The sun and wind as renewable energy sources‏ are attracting more regard as alternative energy sources. In addition to the decreasing fuel sources, also pollution and global warming‏ are important problems. Fuel cells are a beneficial energy technology that generates electric energy through the reaction between hydrogen-rich fuel source and oxygen. The advantage of these devices owing to high efficiency relative to combustion engines, low emissions and producing only heat and water as waste products. Proton exchange membrane fuel cells (PEMFCs) have attracted much interest recently. PEM fuel cells are a high-efficiency power source without pollution for urban vehicles, corroborated by recent legislative initiatives, That's why these devices have become a great deal of attention to the option of fuel-cell-powered vehicles. Previously, some research on modeling and simulation of PEMFC has been performed.[@Secanell2014] The fundamental structure of a PEMFC can be described as two electrodes (anode and cathode) separated by a solid polymer membrane that acts as an electrolyte. Best fuel for fuel cell powered vehicle is hydrogen, this gas causing the highest conversion efficiency for fuel and generating zero tail-pipe emission because the water is an only product of the reaction between hydrogen and air. By flowing Hydrogen fuel through a network of channels to the anode, hydrogen separates into protons that, flow via the membrane to the cathode. An external circuit that linked to the two electrodes containing the collection of electrons as electrical current. Through a similar network of channels the oxygen that comes from the air, named oxidant, flows to the cathode and then, the electrons coming from the external electrical circuit will be received by oxygen and finally, water and heat will be produced from the protons that flowing via the electrolyte membrane.[@Feroldi2011; @Gottesfeld1999]
 
[…]
