### 2018-04-20 718eda3
fix : PEP8 style fixed #37
