+++ opem/Static/Amphlett.py
@@ -3,12 +3,13 @@ import math
 from opem.Params import Amphlett_InputParams as InputParams
 from opem.Params import Amphlett_OutputParams as OutputParams
 from opem.Params import Amphlett_Params_Default as Defaults
-from opem.Params import xi1,xi3,xi4,HHV,uF,Amphlett_Description,Overall_Params_Max_Description,\
-    Overall_Params_Linear_Description,Eth
+from opem.Params import xi1, xi3, xi4, HHV, uF, Amphlett_Description, Overall_Params_Max_Description,\
+    Overall_Params_Linear_Description, Eth
 from opem.Functions import *
 import os
 
-def Power_Thermal_Calc(VStack,N,i):
+
+def Power_Thermal_Calc(VStack, N, i):
     '''
     This function calculate thermal power
     :param VStack: VStack [V]
@@ -20,10 +21,12 @@ def Power_Thermal_Calc(VStack,N,i):
     :return: Thermal power [W]
     '''
     try:
-        return i*((N*Eth)-VStack)
+        return i * ((N * Eth) - VStack)
     except Exception:
         return None
-def Power_Total_Calc(VStack_List,i_step,N):
+
+
+def Power_Total_Calc(VStack_List, i_step, N):
     '''
     This function calculate total elec power and total thermal power by calling integrate function
     :param VStack_List: Vstack list
@@ -35,14 +38,16 @@ def Power_Total_Calc(VStack_List,i_step,N):
     :return: [Total elec power,Total thermal power] as list
     '''
     try:
-        Filtered_List=list(filter(lambda x:x!=None,VStack_List))
-        Filtered_List_Not=list(map(lambda x:(N*Eth)-x,Filtered_List))
-        Total_Elec_Power=integrate(Filtered_List, i_step)
-        Total_Thermal_Power=integrate(Filtered_List_Not,i_step)
-        return [Total_Elec_Power,Total_Thermal_Power]
+        Filtered_List = list(filter(lambda x: x is not None, VStack_List))
+        Filtered_List_Not = list(map(lambda x: (N * Eth) - x, Filtered_List))
+        Total_Elec_Power = integrate(Filtered_List, i_step)
+        Total_Thermal_Power = integrate(Filtered_List_Not, i_step)
+        return [Total_Elec_Power, Total_Thermal_Power]
     except Exception:
-        return [None,None]
-def Linear_Aprox_Params_Calc(B0,B1):
+        return [None, None]
+
+
+def Linear_Aprox_Params_Calc(B0, B1):
     '''
     This function calculate linear approximation overall parameters
     :param B0: intercept
@@ -51,23 +56,24 @@ def Linear_Aprox_Params_Calc(B0,B1):
     :type B1 : float
     :return: [Wmax,Vcell_Wmax] as list
     '''
-    Wmax=0
-    Vcell_Wmax=0
+    Wmax = 0
+    Vcell_Wmax = 0
     try:
-        Wmax=(B0**2)/(4*B1)
+        Wmax = (B0**2) / (4 * B1)
     except Exception:
-        Wmax=None
+        Wmax = None
     try:
-        Vcell_Wmax=(B0/2)
+        Vcell_Wmax = (B0 / 2)
     except Exception:
-        Vcell_Wmax=None
-    if Wmax!=None:
-        Wmax=abs(Wmax)
-    if Vcell_Wmax!=None:
-        Vcell_Wmax=abs(Vcell_Wmax)
-    return [Wmax,Vcell_Wmax]
+        Vcell_Wmax = None
+    if Wmax is not None:
+        Wmax = abs(Wmax)
+    if Vcell_Wmax is not None:
+        Vcell_Wmax = abs(Vcell_Wmax)
+    return [Wmax, Vcell_Wmax]
 
[…]
