+++ opem/Params.py
@@ -1,8 +1,8 @@
 # -*- coding: utf-8 -*-
-Version=0.8
-Website="http://www.ecsim.ir/opem"
-UpdateUrl="http://www.ecsim.ir/opem/update"
-Overview='''
+Version = 0.8
+Website = "http://www.ecsim.ir/opem"
+UpdateUrl = "http://www.ecsim.ir/opem/update"
+Overview = '''
 Modeling and simulation of proton-exchange membrane fuel cells (PEMFC) may work as a
 powerful tool in the Research & development of renewable energy sources. The Open-Source
 PEMFC Simulation Tool (OPEM) is a modeling tool for evaluating the performance of proton
@@ -15,9 +15,9 @@ models of PEMFC that have been proposed in the OPEM, just focus on one particula
 and some others take into account a part or all auxiliaries such as reformers. OPEM is
 a platform for collaborative development of PEMFC models.'''
 
-Warning_Message_1="Warning : The value of I(>{}) leads to minus amount of V, please check your inputs"
-Warning_Message_2="Warning : There are errors in the simulations in some of I amounts; please refer to the .opem file" \
-                  " for review. If you are confident about this parameters, ignore this error."
+Warning_Message_1 = "Warning : The value of I(>{}) leads to minus amount of V, please check your inputs"
+Warning_Message_2 = "Warning : There are errors in the simulations in some of I amounts; please refer to the .opem file" \
+    " for review. If you are confident about this parameters, ignore this error."
 HHV = 1.482
 uF = 0.95
 n = 8 * (10 ** -3)
@@ -25,25 +25,41 @@ m = 3 * (10 ** -5)
 xi1 = -0.948
 xi3 = 7.6 * (10 ** -5)
 xi4 = -1.93 * (10 ** -4)
-#F=96500
-#R1=8.314
-R=8314.47
-F=96484600
-Eth=1.23
-Amphlett_InputParams = {"T": "Cell Operation Temperature [K]", "PH2": "Partial Pressure [atm]", "PO2": "Partial Pressure [atm]",
-               "i-start": "Cell operating current start point [A]", "i-step": "Cell operating current step",
-               "i-stop": "Cell operating current end point [A]", "A": "active area [cm^2]",
-               "l": "Membrane Thickness [cm]",
-               "lambda": "is an adjustable parameter with a min value of 14 and max value of 23",
-               "N": "Number Of Single Cells", "R": "R-Electronic [ohm] (*Optional)",
-               "B": "An empirical constant depending on the cell and its operation state (Tafel Slope) [V]",
-               "JMax": "maximum current density [A/(cm^2)]"}
-Amphlett_OutputParams = {"Enernst": "V", "Eta Activation": "V", "Eta Ohmic": "V", "Eta Concentration": "V", "Loss": "V",
-                "Vcell": "V", "PEM Efficiency": "", "Power": "W", "VStack": "V","Power-Stack":"W","Power-Thermal":"W"}
+# F=96500
+# R1=8.314
+R = 8314.47
+F = 96484600
+Eth = 1.23
+Amphlett_InputParams = {
+    "T": "Cell Operation Temperature [K]",
+    "PH2": "Partial Pressure [atm]",
+    "PO2": "Partial Pressure [atm]",
+    "i-start": "Cell operating current start point [A]",
+    "i-step": "Cell operating current step",
+    "i-stop": "Cell operating current end point [A]",
+    "A": "active area [cm^2]",
[…]
