### 2018-04-20 f340669
fix : get_requirements added to setup file #41
