+++ INSTALL.md
@@ -1,7 +1,7 @@
 ## Installation		
 
 ### Source Code
-- Download [Python3.x](https://www.python.org/downloads/) (>=3.3)
+- Download [Python3.x](https://www.python.org/downloads/) (>=3.4)
 - Download [Version 0.8](https://github.com/ecsim/opem/archive/v0.8.zip) or [Latest Source ](https://github.com/ecsim/opem/archive/master.zip)
 - Run `pip install -r requirements.txt` or `pip3 install -r requirements.txt` (Need root access)
 - Run `python3 setup.py install` or `python setup.py install` (Need root access)				
