### 2018-04-20 8e851e0
fix : import style in examples modified fixed #36
