### 2018-04-20 e32dc12
fix : Amphlett notebook modified fixed #42
