### 2018-04-20 a45bfb8
fix : Larminie_Dicks and Chamberlin-Kim notebook modified fixed #43
