### 2018-04-22 de7d18f
doc : Dynamic notebooks modified fixed #44
