### 2018-04-22 cb69ac5
doc : paper file modified fixed #45
