### 2018-04-22 ECSIM/opem#40 @sepandhaghighi
Hi @nnadeau

It's true, using of general exception is not a good point but in OPEM we don't want to handle different exceptions in different ways so I think it's oky.

Thanks
