### 2018-04-22 ECSIM/opem#38 @sepandhaghighi
@nnadeau Thanks for this point.
I tested `flake8`, warnings are about `from opem.Functions import *`
We need most of functions in opem.Functions for each model so I think it's okay.
