### 2018-04-22 #676 @sepandhaghighi
Hi @nnadeau
Thank you for your time
The issues have been fixed. ;-)
