### 2018-04-24 ECSIM/opem#47 @katyhuff
@kasraaskari : I believe that @nnadeau was not suggesting that you only add the reference to the original source. Rather, I think the current state is still plaigarism as there is no permission from the original author for you to copy it. Indeed, I recommend that you *remove* all text that was plagiarized *and* direct the reader to the original source *instead* of including the copied text.
