### 2018-04-24 ECSIM/opem#41 @katyhuff
@sepandhaghighi It appears 6a9ee30 is failing your unit tests?
