### 2018-04-24 #676 @katyhuff
Thank you for this detailed review @nnadeau .

@sepandhaghighi : all of the issues remain open. Please bring them to closure through conversation with @nnadeau and I. I have left comments on most regarding whether I believe the issues have been sufficiently handled or whether I think the issue looks like it has truly been fixed.

Thanks for working quickly on these issues @sepandhaghighi . I am particularly concerned about the one regarding pem.md. The initial response by your team doesn't seem to recognize the gravity of plaigarism. I hope it will be handled well and swiftly.
