### 2018-04-24 ECSIM/opem#50 @nnadeau
Apply best practices to testing
- Currently, all testing is consolidated in `opem/opem/test.py` as a pseudo-doctest file
  - This is actually improper usage of doctests
  - It is hard to understand how and what is being tested
- If doctests are to be used, they should be implemented in the docstrings
- Otherwise, testing should use a proper framework
- See http://docs.python-guide.org/en/latest/writing/tests/ for best practices
