### 2018-04-24 #676 @nnadeau
@katyhuff @sepandhaghighi I'm still in the process of reviewing the actual codebase.

See https://github.com/ECSIM/opem/issues/50 for my initial observations regarding testing.
