### 2018-04-24 ECSIM/opem#40 @sepandhaghighi
@katyhuff

As I mentioned before, in OPEM we don't want to handle different exceptions in different ways, and there is a single response to all of exceptions so I think general exception is not a problem in this case.

Thanks
