### 2018-04-24 ECSIM/opem#50 @sepandhaghighi
Hi @nnadeau

I used `doctest.testfile`  method because of `.coverage` file merge problem in docstring mode, and we have some output style test in OPEM that I think doctest is better than `pytest` and `unittest` in this case.

Thanks
