### 2018-04-24 #676 @sepandhaghighi
@nnadeau @katyhuff

I merged `develop` and `master` branches and most of issues with `fixed` label closed automatically.

Thanks
