### 2018-04-24 eeb01bf
doc : citation location modified
