+++ paper/paper.md
@@ -39,7 +39,7 @@ bibliography: paper.bib
 						
 
 # Summary
-The sun and wind as renewable energy sources‏ are attracting more regard as alternative energy sources. In addition to the decreasing fuel sources, pollution and global warming‏ are important problems. Fuel cells are a beneficial energy technology that generates electric energy through the reaction between the fuel sources rich in hydrogen and oxygen. In comparsion with combustion engines, fuel cells have many advantages,  such as high efficiency and low emissions. Furthermore the by-products of fuel cells are heat and water. Proton exchange membrane fuel cells (PEMFCs) have attracted much interests recently. PEM fuel cells are Pollution-Free high-efficiency power sources for urban vehicles that recently corporate by legislative initiatives. Previously, some research on modeling and simulation of PEMFC has been performed.[@Secanell2014] The fundamental structure of a PEMFC can be described as two electrodes (anode and cathode) separated by a solid polymer membrane that acting as an electrolyte. The hydrogen gas is the Best fuel for fuel cell powered vehicles, because of the highest conversion efficiency for fuel, generating zero tail-pipe emission and producing water as an only product of the reaction between hydrogen and air .By flowing hydrogen fuel through a network of channels to the anode, hydrogen separates into protons that transfer via the membrane to the cathode. Collection of electrons in the two electrodes causes the creation of electrical current in an electrical circuit that linked to the electrodes. Through a similar network of channels the oxygen that comes from the air, named oxidant, flows to the cathode and then, the electrons coming from the external electrical circuit will be received by oxygen and finally, produce water and heat from the protons that flow via the electrolyte membrane.[@Feroldi2011; @Gottesfeld1999]
+The sun and wind as renewable energy sources‏ are attracting more regard as alternative energy sources. In addition to the decreasing fuel sources, pollution and global warming‏ are important problems. Fuel cells are a beneficial energy technology that generates electric energy through the reaction between the fuel sources rich in hydrogen and oxygen. In comparsion with combustion engines, fuel cells have many advantages,  such as high efficiency and low emissions. Furthermore the by-products of fuel cells are heat and water. Proton exchange membrane fuel cells (PEMFCs) have attracted much interests recently. PEM fuel cells are Pollution-Free high-efficiency power sources for urban vehicles that recently corporate by legislative initiatives. Previously, some research on modeling and simulation of PEMFC has been performed [@Secanell2014]. The fundamental structure of a PEMFC can be described as two electrodes (anode and cathode) separated by a solid polymer membrane that acting as an electrolyte. The hydrogen gas is the Best fuel for fuel cell powered vehicles, because of the highest conversion efficiency for fuel, generating zero tail-pipe emission and producing water as an only product of the reaction between hydrogen and air .By flowing hydrogen fuel through a network of channels to the anode, hydrogen separates into protons that transfer via the membrane to the cathode. Collection of electrons in the two electrodes causes the creation of electrical current in an electrical circuit that linked to the electrodes. Through a similar network of channels the oxygen that comes from the air, named oxidant, flows to the cathode and then, the electrons coming from the external electrical circuit will be received by oxygen and finally, produce water and heat from the protons that flow via the electrolyte membrane [@Feroldi2011; @Gottesfeld1999].
 
[…]
