### 2018-04-25 ECSIM/opem#46 @nnadeau
Please address https://github.com/ECSIM/opem/commit/8d0fad42ade677e644b15db81dbac650f209a992?short_path=e2778e7#commitcomment-28732728
