### 2018-04-25 ECSIM/opem#40 @nnadeau
@sepandhaghighi, unfortunately I disagree. The vast majority of the package's functions have a `try-except` at their returns, catching everything and anything, even for functions that are simply multiplying several floats...

E.g.,
https://github.com/ECSIM/opem/blob/d9748c65ff00ff9f6b730420dc7bb6cad8bc2fca/opem/Static/Amphlett.py#L11-L25

The function is a simple/trivial: `float` * ((`int` * `const float`) - `float`)

There should literally never be an exception...
