### 2018-04-25 ECSIM/opem#50 @nnadeau
Can you please provide a more detailed description of the `merge problem` and `output style test`?
