### 2018-04-25 #676 @nnadeau
@katyhuff @sepandhaghighi

My final review and opinion:

While the package has an `obvious research application`, it does not provide `a significant contribution to the available open source software that either enables some new research challenges to be addressed or makes addressing research challenges significantly better (e.g., faster, easier, simpler)`. (https://joss.theoj.org/about#author_guidelines)

As the majority of function are simple one-liner multiplications, from my perspective, this is more of a `minor, 'utility' package` that would most likely be re-coded by most researchers using `NumPy` and `SciPy`. The most "novel" components are the pretty-print style outputs.

Moreover, the lack of good coding and Python practice makes this difficult for the open-source and research communities to extend and build upon this work. In its current form, this package is best suited for use in a single lab and research group.

@sepandhaghighi, while my overall review may be disappointing, I am very impressed by your team's effort and amazing documentation/README/website. I am also very pleased to see scientific contributions to the Python community. However, from a  journal perspective, I cannot give my stamp of approval.

Best of luck with your future research!
