### 2018-04-25 #676 @sepandhaghighi
@nnadeau

Thanks for your time and deep review
Your points about code style were very good and detailed and we will solve all of them.
But about usage of this package really is not like this, we decided to create this package from some real problems in `PEM` research and focus on models to collect them in one package.
OPEM designed for `PEM` researchers without programming knowledge, it is simple because we collect and solved equations outside of code.
We spent about 10 months to collect and validate these models, now is really easy to re-code but this procedure is not.
There is no package like OPEM that support all of these models in this style and I think we provide easier and simpler way in `PEM` analysis and we are in `JOSS` scope.

Thanks ;-)
