### 2018-04-25 ECSIM/opem#50 @sepandhaghighi
@nnadeau

Merge problem : each function tests in `docstring` mode create one `.coverage` file and this file overwrite by next.

Output style test : some functions output style in console

Thanks
