+++ opem/Test/Padulles_Amphlett_test.py
@@ -1,0 +1,565 @@
+# -*- coding: utf-8 -*-
+'''
+>>> from opem.Dynamic.Padulles_Amphlett import *
+>>> Test_Vector={"A":50.6,"l":0.0178,"lambda":23,"JMax":1.5,"T":343,"N0":5,"KO2":0.0000211,"KH2":0.0000422,"KH2O":0.000007716,"tH2":3.37,"tO2":6.74,"t1":2,"t2":2,"tH2O":18.418,"B":0.016,"rho":1.168,"qMethanol":0.0002,"CV":2,"i-start":0.1,"i-stop":4,"i-step":0.1,"Name":"Test"}
+>>> Padulles_Amphlett_Data=Dynamic_Analysis(InputMethod=Test_Vector, TestMode=True)
+###########
+Padulles-Amphlett-Model Simulation
+###########
+Analyzing . . .
+I : 0.1
+E : 6.0684154992732005 V
+Eta Activation : 0.18557231242539243 V
+Eta Concentration : 2.109426805213159e-05 V
+Eta Ohmic : 0.00017548304819292376 V
+FC Efficiency : 0.6589193654570529
+FC Power : 0.5139571050565013 W
+FC Voltage : 5.139571050565013 V
+Loss : 0.1857688897416375 V
+PH2 : 0.19717074233280188 atm
+PH2O : 0.2426831613626925 atm
+PO2 : 0.1906263686382979 atm
+Power-Thermal : 0.10104289494349877 W
+###########
+I : 0.2
+E : 6.068413961701556 V
+Eta Activation : 0.23146009851376736 V
+Eta Concentration : 4.221638333089875e-05 V
+Eta Ohmic : 0.0003510800160998837 V
+FC Efficiency : 0.629377818863534
+FC Power : 0.9818293974271133 W
+FC Voltage : 4.909146987135566 V
+Loss : 0.23185339491319815 V
+PH2 : 0.1971566919511875 atm
+PH2O : 0.24266586776736396 atm
+PO2 : 0.1906184358000996 atm
+Power-Thermal : 0.2481706025728869 W
+###########
+I : 0.3
+E : 6.068412424065923 V
+Eta Activation : 0.2583036192079603 V
+Eta Concentration : 6.336641945755048e-05 V
+Eta Ohmic : 0.0005267910327125488 V
+FC Efficiency : 0.6120440436878553
+FC Power : 1.4321830622295815 W
+FC Voltage : 4.773943540765272 V
+Loss : 0.25889377666013036 V
+PH2 : 0.19714264156957312 atm
+PH2O : 0.24264857417203542 atm
+PO2 : 0.1906105029619013 atm
+Power-Thermal : 0.41281693777041856 W
+###########
+I : 0.4
+E : 6.068410886366294 V
+Eta Activation : 0.27735002084480426 V
+Eta Concentration : 8.454445034568427e-05 V
+Eta Ohmic : 0.0007026162388380664 V
+FC Efficiency : 0.5997083306020967
+FC Power : 1.871089991478542 W
+FC Voltage : 4.677724978696355 V
+Loss : 0.278137181533988 V
+PH2 : 0.19712859118795872 atm
+PH2O : 0.24263128057670688 atm
+PO2 : 0.19060257012370302 atm
+Power-Thermal : 0.5889100085214584 W
+###########
+I : 0.5
+E : 6.068409348602667 V
+Eta Activation : 0.2921240370409447 V
+Eta Concentration : 0.00010575055020278165 V
+Eta Ohmic : 0.0008785557847524419 V
+FC Efficiency : 0.5901112348363035
+FC Power : 2.3014338158615835 W
+FC Voltage : 4.602867631723167 V
+Loss : 0.2931083433758999 V
[…]
