+++ setup.py
@@ -13,7 +13,7 @@ def get_requirements():
 
 setup(
   name = 'opem',
-  packages = ['opem','opem.Static','opem.Dynamic'],
+  packages = ['opem','opem.Static','opem.Dynamic','opem.Test'],
   version = '0.8',
   description = 'Open Source PEM Cell Simulation Tool',
   long_description='''

[…]
