### 2018-04-26 3f224a9
fix : import * in Functions.py modified fixed #38
