+++ opem/Static/Chamberline_Kim.py
@@ -3,7 +3,7 @@ import math
 from opem.Params import Chamberline_InputParams as InputParams
 from opem.Params import Chamberline_OutputParams as OutputParams
 from opem.Static.Amphlett import Efficiency_Calc, Power_Calc, VStack_Calc, PowerStack_Calc, Power_Thermal_Calc, Power_Total_Calc, Linear_Aprox_Params_Calc, Max_Params_Calc
-from opem.Functions import *
+import opem.Functions
 from opem.Params import Chamberline_Description, Overall_Params_Max_Description, Overall_Params_Linear_Description
 import os
 import datetime
@@ -40,7 +40,7 @@ def Vcell_Calc(E0, b, R, m, n, i, A):
 
 
 def Static_Analysis(
-        InputMethod=Get_Input,
+        InputMethod=opem.Functions.Get_Input,
         TestMode=False,
         PrintMode=True,
         ReportMode=True):
@@ -81,16 +81,17 @@ def Static_Analysis(
             print("Analyzing . . .")
         Name = Input_Dict["Name"]
         if ReportMode:
-            OutputFile = Output_Init(Input_Dict, Simulation_Title, Name)
-            CSVFile = CSV_Init(
+            OutputFile = opem.Functions.Output_Init(
+                Input_Dict, Simulation_Title, Name)
+            CSVFile = opem.Functions.CSV_Init(
                 OutputParamsKeys,
                 OutputParams,
                 Simulation_Title,
                 Name)
-            HTMLFile = HTML_Init(Simulation_Title, Name)
+            HTMLFile = opem.Functions.HTML_Init(Simulation_Title, Name)
         IEnd = Input_Dict["i-stop"]
         IStep = Input_Dict["i-step"]
-        Precision = get_precision(IStep)
+        Precision = opem.Functions.get_precision(IStep)
         i = Input_Dict["i-start"]
         I_List = []
         Efficiency_List = []
@@ -108,9 +109,9 @@ def Static_Analysis(
                     Input_Dict["n"],
                     i,
                     Input_Dict["A"])
-                [Warning1, I_Warning] = warning_check_1(
+                [Warning1, I_Warning] = opem.Functions.warning_check_1(
                     Output_Dict["Vcell"], I_Warning, i, Warning1)
-                Warning2 = warning_check_2(
+                Warning2 = opem.Functions.warning_check_2(
                     Vcell=Output_Dict["Vcell"], warning_flag=Warning2)
                 Output_Dict["PEM Efficiency"] = Efficiency_Calc(
                     Output_Dict["Vcell"])
@@ -126,28 +127,31 @@ def Static_Analysis(
                 Power_List.append(Output_Dict["Power-Stack"])
                 Power_Thermal_List.append(Output_Dict["Power-Thermal"])
                 if ReportMode:
-                    Output_Save(
+                    opem.Functions.Output_Save(
                         OutputParamsKeys,
                         Output_Dict,
                         OutputParams,
                         i,
                         OutputFile,
                         PrintMode)
-                    CSV_Save(OutputParamsKeys, Output_Dict, i, CSVFile)
-                i = rounder(i + IStep, Precision)
+                    opem.Functions.CSV_Save(
+    OutputParamsKeys, Output_Dict, i, CSVFile)
+                i = opem.Functions.rounder(i + IStep, Precision)
             except Exception as e:
                 print(e)
-                i = rounder(i + IStep, Precision)
+                i = opem.Functions.rounder(i + IStep, Precision)
                 if ReportMode:
-                    Output_Save(
+                    opem.Functions.Output_Save(
                         OutputParamsKeys,
                         Output_Dict,
                         OutputParams,
[…]
