+++ opem/Dynamic/Padulles1.py
@@ -4,7 +4,7 @@ from opem.Params import Padulles_InputParams as InputParams
 from opem.Params import Padulles_Outparams as OutputParams
 from opem.Params import R, F, uF, HHV, Padulles_Description, Overall_Params_Max_Description, Overall_Params_Linear_Description
 from opem.Static.Amphlett import Power_Calc, Power_Thermal_Calc, Power_Total_Calc, Linear_Aprox_Params_Calc, Max_Params_Calc
-from opem.Functions import *
+import opem.Functions
 import os
 
 
@@ -155,7 +155,7 @@ def Efficiency_Calc(Vcell, N):
 
 
 def Dynamic_Analysis(
-        InputMethod=Get_Input,
+        InputMethod=opem.Functions.Get_Input,
         TestMode=False,
         PrintMode=True,
         ReportMode=True):
@@ -196,16 +196,17 @@ def Dynamic_Analysis(
             print("Analyzing . . .")
         Name = Input_Dict["Name"]
         if ReportMode:
-            OutputFile = Output_Init(Input_Dict, Simulation_Title, Name)
-            CSVFile = CSV_Init(
+            OutputFile = opem.Functions.Output_Init(
+                Input_Dict, Simulation_Title, Name)
+            CSVFile = opem.Functions.CSV_Init(
                 OutputParamsKeys,
                 OutputParams,
                 Simulation_Title,
                 Name)
-            HTMLFile = HTML_Init(Simulation_Title, Name)
+            HTMLFile = opem.Functions.HTML_Init(Simulation_Title, Name)
         IEnd = Input_Dict["i-stop"]
         IStep = Input_Dict["i-step"]
-        Precision = get_precision(IStep)
+        Precision = opem.Functions.get_precision(IStep)
         i = Input_Dict["i-start"]
         I_List = []
         Power_List = []
@@ -233,9 +234,9 @@ def Dynamic_Analysis(
                     Output_Dict["PO2"])
                 Output_Dict["FC Voltage"] = Vcell_Calc(
                     Output_Dict["E"], Input_Dict["B"], Input_Dict["C"], i, Input_Dict["Rint"])
-                [Warning1, I_Warning] = warning_check_1(
+                [Warning1, I_Warning] = opem.Functions.warning_check_1(
                     Output_Dict["FC Voltage"], I_Warning, i, Warning1)
-                Warning2 = warning_check_2(
+                Warning2 = opem.Functions.warning_check_2(
                     Vcell=Output_Dict["FC Voltage"],
                     warning_flag=Warning2)
                 Vstack_List.append(Output_Dict["FC Voltage"])
@@ -249,28 +250,31 @@ def Dynamic_Analysis(
                 Power_List.append(Output_Dict["FC Power"])
                 Power_Thermal_List.append(Output_Dict["Power-Thermal"])
                 if ReportMode:
-                    Output_Save(
+                    opem.Functions.Output_Save(
                         OutputParamsKeys,
                         Output_Dict,
                         OutputParams,
                         i,
                         OutputFile,
                         PrintMode)
-                    CSV_Save(OutputParamsKeys, Output_Dict, i, CSVFile)
-                i = rounder(i + IStep, Precision)
+                    opem.Functions.CSV_Save(
+    OutputParamsKeys, Output_Dict, i, CSVFile)
+                i = opem.Functions.rounder(i + IStep, Precision)
             except Exception as e:
                 print(str(e))
-                i = rounder(i + IStep, Precision)
+                i = opem.Functions.rounder(i + IStep, Precision)
                 if ReportMode:
-                    Output_Save(
+                    opem.Functions.Output_Save(
                         OutputParamsKeys,
[…]

[…]
