### 2018-04-27 e3d18f2
fix : __init__ file docstring added
