+++ opem/__main__.py
@@ -1,5 +1,11 @@
 # -*- coding: utf-8 -*-
-
+'''
+OPEM Main File
+'''
+import doctest
+import sys
+import coverage
+import os
 from opem.Static.Amphlett import Static_Analysis as Amphlett_Analysis
 from opem.Static.Larminie_Dicks import Static_Analysis as Larminiee_Analysis
 from opem.Static.Chamberline_Kim import Static_Analysis as Chamberline_Kim_Analysis
@@ -10,36 +16,39 @@ from opem.Dynamic.Padulles_Amphlett import Dynamic_Analysis as Padulles_Amphlett
 from art import tprint
 from opem.Params import Version, Description_Menu, Description_Links, Vectors
 from opem.Functions import check_update, description_print, description_control
-import doctest
-import sys
-import coverage
-import os
 
-def run_test(Test_Path):
+
+def run_test(test_path):
+    '''
+    This function run test cases
+    :param test_path: test files path
+    :type test_path : str
+    :return: None
+    '''
     cov = coverage.Coverage()
     cov.start()
-    doctest.testfile(os.path.join(Test_Path, "Amphlett_test.py"),
+    doctest.testfile(os.path.join(test_path, "Amphlett_test.py"),
                      optionflags=doctest.NORMALIZE_WHITESPACE | doctest.ELLIPSIS,
                      verbose=False)
-    doctest.testfile(os.path.join(Test_Path, "Chamberline_Kim_test.py"),
+    doctest.testfile(os.path.join(test_path, "Chamberline_Kim_test.py"),
                      optionflags=doctest.NORMALIZE_WHITESPACE | doctest.ELLIPSIS,
                      verbose=False)
-    doctest.testfile(os.path.join(Test_Path, "Functions_test.py"),
+    doctest.testfile(os.path.join(test_path, "Functions_test.py"),
                      optionflags=doctest.NORMALIZE_WHITESPACE | doctest.ELLIPSIS,
                      verbose=False)
-    doctest.testfile(os.path.join(Test_Path, "Larminie_Dicks_test.py"),
+    doctest.testfile(os.path.join(test_path, "Larminie_Dicks_test.py"),
                      optionflags=doctest.NORMALIZE_WHITESPACE | doctest.ELLIPSIS,
                      verbose=False)
-    doctest.testfile(os.path.join(Test_Path, "Padulles_Amphlett_test.py"),
+    doctest.testfile(os.path.join(test_path, "Padulles_Amphlett_test.py"),
                      optionflags=doctest.NORMALIZE_WHITESPACE | doctest.ELLIPSIS,
                      verbose=False)
-    doctest.testfile(os.path.join(Test_Path, "Padulles_Hauer_test.py"),
+    doctest.testfile(os.path.join(test_path, "Padulles_Hauer_test.py"),
                      optionflags=doctest.NORMALIZE_WHITESPACE | doctest.ELLIPSIS,
                      verbose=False)
-    doctest.testfile(os.path.join(Test_Path, "Padulles1_test.py"),
+    doctest.testfile(os.path.join(test_path, "Padulles1_test.py"),
                      optionflags=doctest.NORMALIZE_WHITESPACE | doctest.ELLIPSIS,
                      verbose=False)
-    doctest.testfile(os.path.join(Test_Path, "Padulles2_test.py"),
+    doctest.testfile(os.path.join(test_path, "Padulles2_test.py"),
                      optionflags=doctest.NORMALIZE_WHITESPACE | doctest.ELLIPSIS,
                      verbose=False)
     cov.stop()
@@ -47,9 +56,9 @@ def run_test(Test_Path):
     cov.save()
 
[…]
