+++ opem/Dynamic/Padulles2.py
@@ -170,7 +170,7 @@ def Dynamic_Analysis(
                         OutputFile,
                         PrintMode)
                     opem.Functions.CSV_Save(
-    OutputParamsKeys, Output_Dict, i, CSVFile)
+                        OutputParamsKeys, Output_Dict, i, CSVFile)
                 i = opem.Functions.rounder(i + IStep, Precision)
             except Exception as e:
                 print(str(e))
@@ -184,7 +184,7 @@ def Dynamic_Analysis(
                         OutputFile,
                         PrintMode)
                     opem.Functions.CSV_Save(
-    OutputParamsKeys, Output_Dict, i, CSVFile)
+                        OutputParamsKeys, Output_Dict, i, CSVFile)
         [Estimated_V, B0, B1] = opem.Functions.linear_plot(
             x=I_List, y=Vstack_List)
         Linear_Approx_Params = Linear_Aprox_Params_Calc(B0, B1)
@@ -202,7 +202,7 @@ def Dynamic_Analysis(
         Overall_Params_Max["Ptotal(Thermal)"] = Power_Total[1]
         if ReportMode:
             opem.Functions.HTML_Desc(
-    Simulation_Title, Padulles2_Description, HTMLFile)
+                Simulation_Title, Padulles2_Description, HTMLFile)
             opem.Functions.HTML_Input_Table(
                 Input_Dict=Input_Dict,
                 Input_Params=InputParams,
@@ -268,14 +268,14 @@ def Dynamic_Analysis(
                 size="600px",
                 file=HTMLFile)
             opem.Functions.HTML_Chart(x=str(list(map(opem.Functions.rounder,
-                                      Power_List))),
-                       y=str(Efficiency_List),
-                       color='rgb(238, 210, 141)',
-                       x_label="P(W)",
-                       y_label="EFF",
-                       chart_name="Efficiency vs Power",
-                       size="600px",
-                       file=HTMLFile)
+                                                     Power_List))),
+                                      y=str(Efficiency_List),
+                                      color='rgb(238, 210, 141)',
+                                      x_label="P(W)",
+                                      y_label="EFF",
+                                      chart_name="Efficiency vs Power",
+                                      size="600px",
+                                      file=HTMLFile)
             opem.Functions.HTML_Chart(
                 x=str(I_List),
                 y=str(Power_Thermal_List),
