### 2018-04-28 81c9d58
fix : general exception removed fixed #40
