### 2018-04-28 5ff3ec0
fix : check_update function exception error modified
