### 2018-04-28 743ac86
fix : check_update test case added
