+++ opem/Test/Functions_test.py
@@ -39,5 +39,11 @@ Test
 <BLANKLINE>
 <BLANKLINE>
 ###########
+>>> check_update(1)
+>>> check_update(0.1)
+###########
+New Version (0.8) Is Available!
+Website : http://www.ecsim.ir/opem
+###########
 
 '''
