### 2018-04-28 f1d4707
fix : HTML_Chart_Multi function removed
