+++ opem/Functions.py
@@ -428,59 +428,6 @@ def HTML_Chart(x, y, color, x_label, y_label, chart_name, size, file):
             chart_data))
 
 
-def HTML_Chart_Multi(
-        x,
-        y,
-        color,
-        x_label,
-        y_label,
-        chart_name,
-        size,
-        axes,
-        file):
-    '''
-    This function write chartjs chart in html file
-    :param x: x data as a string list
-    :type x : str
-    :param y: y data as string list (or list of y)
-    :param color: color code of chart (or list of color)
-    :param x_label:x-axis label
-    :type x_label : str
-    :param y_label:y-axis labels
-    :type y_label : list
-    :param chart_name: chart name (or list of chart_name)
-    :param size: chart size in pixel
-    :type size : str
-    :param file: html file object
-    :type file : file object
-    :return: None
-    '''
-    chart_data = ""
-    chart_title = str(chart_name)
-    if isinstance(y, list):
-        y_data = list(map(None_Omit, y))
-        for index, data in enumerate(y_data):
-            chart_data += opem.Script.MULTI_AXES_CHART_DATA.format(
-                chart_name[index], data, color[index], axes[index])
-            if index != len(y_data) - 1:
-                chart_data += ","
-            chart_data += "\n"
-    else:
-        y_data = None_Omit(y)
-        chart_data = opem.Script.MULTI_AXES_CHART_DATA.format(
-            chart_name, y_data, color, axes)
-    x_data = None_Omit(x)
-    file.write(
-        opem.Script.MULTI_AXES_LINE_CHART.format(
-            x_data,
-            y_label[0],
-            y_label[1],
-            x_label,
-            chart_title,
-            size,
-            chart_data))
-
-
 def HTML_Input_Table(Input_Dict, Input_Params, file):
     '''
     This function add table to html file
