+++ opem/Static/Chamberline_Kim.py
@@ -31,7 +31,7 @@ def Vcell_Calc(E0, b, R, m, n, i, A):
         J = i / A
         result = E0 - b * math.log(J) - R * J - m * math.exp(n * J)
         return result
-    except (TypeError,OverflowError,ValueError):
+    except (TypeError, OverflowError, ValueError):
         print(
             "[Error] Vcell Calculation Error (E0:%s, b:%s, R:%s, m:%s, n:%s, i:%s, A:%s)" %
             (str(E0), str(b), str(R), str(m), str(n), str(i), str(A)))
