+++ requirements.txt
@@ -1,3 +1,5 @@
 codecov==2.0.15
 art>0.7
 requests==2.18.4
+pytest
+pytest-cov
