+++ .travis.yml
@@ -12,6 +12,6 @@ install:
   - python setup.py install
 
 script:
-  - python -m opem test
+  - python -m pytest --cov=opem --cov-report=term
 after_success:
   - codecov
