+++ .travis.yml
@@ -12,6 +12,6 @@ install:
   - python setup.py install
 
 script:
-  - python -m pytest --cov=opem --cov-report=term
+  - python -m pytest opem/Test --cov=opem --cov-report=term
 after_success:
   - codecov
