### 2018-04-28 bb52c5e
fix : pytest and pytest-cov version pinned fixed #50
