### 2018-04-29 01bc0e1
fix : MANIFEST file updated
