+++ MANIFEST.in
@@ -3,6 +3,7 @@ include *.md
 include *.spec
 include *.txt
 include *.yml
+include *.ini
 recursive-include Documents *.ipynb
 recursive-include Documents *.jpg
 recursive-include Documents *.md
