+++ opem/test.py
@@ -1,3000 +1,0 @@
-# -*- coding: utf-8 -*-
-'''
->>> import coverage
->>> from opem.Static.Amphlett import *
->>> import random
->>> cov=coverage.Coverage()
->>> cov.start()
->>> data=[i for i in range(100)]
->>> integrate(data,1)
-4867.666666666666
->>> data[0]=None
->>> integrate(data,1)
->>> linear_plot([1,2,3],[2,4,6])
-[[2.0, 4.0, 6.0], 0.0, 2.0]
->>> isfloat("2")
-True
->>> isfloat("2.02")
-True
->>> isfloat('ss')
-False
->>> filter_lambda({"lambda":24})
-[Warning] Opem Automatically Set Lambda To Maximum Value (23)
-{'lambda': 23}
->>> filter_alpha({"alpha":2})
-[Warning] Opem Automatically Set Alpha To Maximum Value (1)
-{'alpha': 1}
->>> filter_lambda({"lambda":13})
-[Warning] Opem Automatically Set Lambda To Minimum Value (14)
-{'lambda': 23}
->>> filter_alpha({"alpha":-0.1})
-[Warning] Opem Automatically Set Alpha To Maximum Value (0)
-{'alpha': 0}
->>> T=343.15
->>> PH2=1
->>> PO2=1
->>> i=0
->>> A=50.6
->>> l=0.178
->>> lambda_param=23
->>> N=1
->>> Jn=0.003
->>> JMax=0.469
->>> Enernst_Calc(T,PH2,PO2)
-1.19075
->>> CH2_Calc(PH2,T)
-7.330294784824117e-07
->>> CO2_Calc(PO2,T)
-8.402541445801334e-07
->>> Rho_Calc(i,A,T,lambda_param)
-4.978789826264977
->>> Xi2_Calc(A,PH2,T)
-0.0030373688787134006
->>> Eta_Conc_Calc(i,A,Jn,JMax)
-0
->>> Eta_Ohmic_Calc(i,l,A,T,lambda_param)
-0
->>> Eta_Act_Calc(T,PO2,PH2,i,A)
-0
->>> PowerStack_Calc(2,2)
-4
->>> PowerStack_Calc(None,2)
-[Error] Power Stack Calculation Error (Power:None, N:2)
->>> T='20000000000'
->>> PH2='10000000'
->>> PO2='1000000000'
->>> i='160000000'
->>> A='30000000000'
->>> l='50000000000'
->>> lambda_param='50000000000'
->>> N='80000000000'
->>> Enernst_Calc(T,PH2,PO2)
-[Error] Enernst Calculation Failed (T:20000000000 , PH2:10000000, PO2:1000000000)
->>> CH2_Calc(PH2,T)
-[Error] CH2 Calculation Failed (PH2:10000000, T:20000000000)
->>> CO2_Calc(PO2,T)
-[Error] CO2 Calculation Failed (PO2:1000000000, T:20000000000)
->>> Rho_Calc(i,A,T,lambda_param)
-[Error] Rho Calculation Failed (i:160000000, A:30000000000, T:20000000000, lambda:50000000000)
->>> Xi2_Calc(A,PH2,T)
-[Error] CH2 Calculation Failed (PH2:10000000, T:20000000000)
-[Error] Xi2 Calculation Failed (A:30000000000, PH2:10000000, T:20000000000)
->>> Eta_Conc_Calc(i,A,None,None)
-[Error] Eta Concentration Calculation Failed (i:160000000, A:30000000000, B:None, JMax:None)
->>> Eta_Ohmic_Calc(i,l,A,T,lambda_param)
-[Error] Rho Calculation Failed (i:160000000, A:30000000000, T:20000000000, lambda:50000000000)
[…]
