### 2018-05-03 ff43151
Add license scan report and status
Signed-off-by: fossabot <badges@fossa.io>
