### 2018-05-23 d23f0f4
Update pytest from 3.5.1 to 3.6.0
