### 2018-06-05 4aab54c
Update pytest from 3.5.1 to 3.6.1
