### 2018-06-20 aa7403e
Update pytest from 3.6.1 to 3.6.2
