### 2018-07-05 6edfc85
Update pytest from 3.6.2 to 3.6.3
