### 2018-07-11 #676 @katyhuff
Hi Folks, I'm sorry for the delay here. It took some time and exploration to determine whether there is really a fit concern for JOSS. @nnadeau , I thank you for your detailed and careful review. I understand your concern about this potentially being a minor utility rather than a tool enabling new research. However, after investigating a bit, I personally evaluate the work more generously. While it may only enable new research in this small subdomain, and perhaps only in this research group, I am convinced that it is more than a minor utility and will indeed, enable new research. I would like to approve this paper.

@sepandhaghighi @kasraaskari At this point could you make an archive of the reviewed software in Zenodo/figshare/other service and update this thread with the DOI of the archive? I can then move forward with accepting the submission.
