### 2018-07-12 595122c
doc : CHANGELOG updated
