### 2018-07-12 beb20cf
doc : CHANGELOG updated
