### 2018-07-13 781ebed
fix : get_requirements function updated
