+++ setup.py
@@ -10,12 +10,18 @@ def get_requirements():
     requirements = open("requirements.txt", "r").read()
     return list(filter(lambda x: x != "", requirements.split()))
 
-setup(
-    name='opem',
-    packages=['opem', 'opem.Static', 'opem.Dynamic', 'opem.Test'],
-    version='0.9',
-    description='Open Source PEM Cell Simulation Tool',
-    long_description='''
+
+def read_description():
+    try:
+        with open("README.md") as r:
+            description = "\n"
+            description += r.read()
+        with open("CHANGELOG.md") as c:
+            description += "\n"
+            description += c.read()
+        return description
+    except Exception:
+        return '''
   Modeling and simulation of proton-exchange membrane fuel cells (PEMFC) may work as a powerful tool in the Research &
   development of renewable energy sources. The Open-Source PEMFC Simulation Tool (OPEM) is a modeling tool for
   evaluating the performance of proton exchange membrane fuel cells. This package is a combination of models
@@ -24,7 +30,17 @@ setup(
   and compositions, cell temperature and current density, but also cell parameters including the active area and
   membrane thickness. In addition, some of the different models of PEMFC that have been proposed in the OPEM,
   just focus on one particular FC stack, and some others take into account a part or all auxiliaries such as
-  reformers. OPEM is a platform for collaborative development of PEMFC models.''',
+  reformers. OPEM is a platform for collaborative development of PEMFC models.'''
+
+
+
+setup(
+    name='opem',
+    packages=['opem', 'opem.Static', 'opem.Dynamic', 'opem.Test'],
+    version='0.9',
+    description='Open Source PEM Cell Simulation Tool',
+    long_description=read_description(),
+    long_description_content_type='text/markdown',
     author='Sepand Haghighi,Kasra Askari,Sarmin Hamidi,Mohammad Mahdi Rahimi',
     author_email='opem@ecsim.ir',
     url='https://github.com/ecsim/opem',
