+++ appveyor.yml
@@ -2,6 +2,13 @@ build: false
 
 environment:
   matrix:
+  - PYTHON: "C:\\Python33"
+    PYTHON_VERSION: "3.3.0"
+    PYTHON_ARCH: "64"
+    
+  - PYTHON: "C:\\Python33"
+    PYTHON_VERSION: "3.3.0"
+    PYTHON_ARCH: "32"
     
   - PYTHON: "C:\\Python34"
     PYTHON_VERSION: "3.4.0"
