+++ .travis.yml
@@ -6,6 +6,7 @@ python:
   - 3.6
   - 3.5
   - 3.4
+  - 3.3
 
 install:
   - pip install -r requirements.txt
