### 2018-07-13 f95cd54
doc : README links convert to global
