### 2018-07-13 38ca769
fix : some links changed to global in README
