### 2018-07-13 b4e5695
doc : images global links updated
