### 2018-07-13 1e44e4b
doc : CHANGELOG updated
