### 2018-07-13 45fdc90
doc : INSTALL EXE link updated
